refactor(framework): extract pipeline executor from route.ts
Moves runSteps (the step queue loop) and processError verbatim into
pipeline/executor.ts as runPipeline(deps, exchange, startTime) with an
explicit ExecutorDeps object; only this.* references became deps.*, so
event emission order is unchanged (verified by the telemetry ordering
suite). route.ts now holds lifecycle, consumer wiring, exchange
construction, and forwarding: 1,870 -> 838 lines across the B3 series.
